# BE-swagger.yaml
openapi: 3.0.4
info:
  title: Blood Donation Support System - OpenAPI 1.0
  version: v1
servers:
  - url: '/'
  - url: http://localhost:8080
  - url: https://be-t8i8.onrender.com
...
paths:
...
  /api/users/change-password:
    post:
      tags:
        - Users
      summary: Change Password
      description: Change password
      operationId: changePassword
      requestBody:
        description: Information change password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordReqBody'
      responses:
        default:
          description: Change password success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
  /api/users/is-active/{user_id}:
    patch:
      tags:
        - Users
      summary: Change is active user for Admin
      description: Change is active user for Admin
      operationId: changeIsActive
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: MongoId
      responses:
        default:
          description: Change is active success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
...
  /api/users/role-for-admin/{user_id}:
    patch:
      tags:
        - Users
      summary: Change role user for Admin
      description: Change role user for Admin
      operationId: changeRoleForAdmin
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: MongoId
      requestBody:
        description: Information to change user role
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  $ref: '#/components/schemas/UserRole'
              required:
                - role
      responses:
        default:
          description: Change role success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthentication'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
...
  /api/requests/request-registrations:
    post:
      tags:
        - Request Registrations
      summary: Create Request Registration
      description: Create a new request registration
      operationId: createRequestRegistration
      requestBody:
        description: Information create request registration
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestRegistrationReqBody'
        required: true
      responses:
        '200':
          description: Create donation registration success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestRegistration'
        '422':
          description: Invalid Input
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []

    get:
      tags:
        - Request Registrations
      summary: Get All Request Registrations for Staff or Admin
      description: G Get All Request Registrations for Staff or Admin
      operationId: getAllRequestRegistrations
      responses:
        default:
          description: Get all request registrations success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
  /api/requests/request-registrations/user:
    get:
      tags:
        - Request Registrations
      summary: Get Request Registration by User ID Now
      description: Get Request Registration by User ID Now
      operationId: getRequestRegistrationByUserId
      responses:
        default:
          description: Get request registration by user id success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
  /api/requests/request-registrations/{id}:
    get:
      tags:
        - Request Registrations
      summary: Get Request Registrations By Id
      description: Get Request Registrations By Id
      operationId: getRequestRegistrationById
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: MongoId
          required: true
      responses:
        default:
          description: Get request registration success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
    patch:
      tags:
        - Request Registrations
      summary: Update Request Registrations By Id for Staff or Admin
      description: Update Request Registrations By Id for Staff or Admin
      operationId: updateRequestRegistration
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: MongoId
          required: true
      requestBody:
        description: Information update donation processes
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestRegistrationReqBody'
        required: true
      responses:
        default:
          description: Update donation processes success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestRegistration'
        '401':
          description: UnAuthorization
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    ChangePasswordReqBody:
      type: object
      properties:
        old_password:
          type: string
          example: Duc@123123
        password:
          type: string
          example: Duc@123123
        confirm_password:
          type: string
          example: Duc@123123
...
    User:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        full_name:
          type: string
          example: Dustin Tsan
        email:
          type: string
          example: dustintsan1@gmail.com
        date_of_birth:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        role:
          $ref: '#/components/schemas/UserRole'
        gender:
          $ref: '#/components/schemas/UserGender'
        blood_group_id:
          type: string
          nullable: true
          example: null
        number_of_donation:
          type: integer
          example: 0
        number_of_request:
          type: integer
          example: 0
        weight:
          type: number
          example: 0
        location:
          type: string
          example: ''
        phone:
          type: string
          example: ''
        avatar_url:
          type: string
          example: ''
        created_at:
          type: string
          format: date-time
          example: 2025-05-22T18:17:35.277Z
        updated_at:
          type: string
          format: date-time
          example: 2025-05-22T18:17:35.277Z
    SuccessAuthentication:
      type: object
      properties:
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            access_token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjgyZjdiZTM1N2FlNzVhMGI5NmI3YWM1IiwidG9rZW5fdHlwZSI6MCwicm9sZSI6IlN0YWZmIiwiaWF0IjoxNzQ3OTQ0ODk0LCJleHAiOjE3NDc5NDU3OTR9.WpOyu0HlgqU2fg7ElIakw-JVM5TWIUaVb-syRrgovOI
            refresh_token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjgyZjdiZTM1N2FlNzVhMGI5NmI3YWM1IiwidG9rZW5fdHlwZSI6MCwicm9sZSI6IlN0YWZmIiwiaWF0IjoxNzQ3OTQ0ODk0LCJleHAiOjE3NDc5NDU3OTR9.WpOyu0HlgqU2fg7ElIakw-JVM5TWIUaVb-syRrgovOI
    UserGender:
      type: string
      enum: [Other, Male, Female]
      example: Male
    UserRole:
      type: string
      enum: [Admin, Staff, Customer]
      example: Admin
...
    RequestType:
      type: string
      enum:
        - Whole Blood # Hiến máu toàn phần
        - Platelets # Hiến tiểu cầu
        - Plasma # Hiến huyết tương
        - Red Blood Cells # Hiến hồng cầu kép
        - Platelets - Plasma # Hiến tiểu cầu + huyết tương
        - Plasma - Red Blood Cells # Hiến huyết tương + hồng cầu
        - Platelets - Red Blood Cells # Hiến tiểu cầu + hồng cầu
      example: Whole Blood
...
    RequestRegistrationStatus:
      type: string
      enum: [Pending, Approved, Rejected]
      example: Pending
    CreateRequestRegistrationReqBody:
      type: object
      properties:
        blood_group_id:
          type: string
          format: MongoId
        request_type:
          $ref: '#/components/schemas/RequestType'
        receive_date_request:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        citizen_id_number:
          type: string
          example: 123123123123
        is_emergency:
          type: boolean
          example: true
        full_name:
          type: string
        phone:
          type: string
        image:
          type: string
        note:
          type: string
    RequestRegistration:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        user_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        blood_group_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        blood_group_name:
          type: string
          example: 'B+'
        citizen_id_number:
          type: string
          example: 123123123123
        phone:
          type: string
          example: 0987654321
        full_name:
          type: string
          example: 'Thao Nguyen'
        blood_component_ids:
          type: array
          format: MongoId
          example: [682f6a3fdb3fb589d23e7654, 682f6a3fdb3fb589d23e7654]
        request_process_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        health_check_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        receive_date_request:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        request_type:
          $ref: '#/components/schemas/RequestType'
        status:
          $ref: '#/components/schemas/RequestRegistrationStatus'
        is_emergency:
          type: boolean
          example: true
        updated_by:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        note:
          type: string
        image:
          type: string
        created_at:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        updated_at:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
    RequestRegistrationReqBody:
      type: object
      properties:
        blood_group_id:
          type: string
          format: MongoId
          example: 682f6a3fdb3fb589d23e7654
        request_type:
          $ref: '#/components/schemas/RequestType'
        image:
          type: string
        is_emergency:
          type: boolean
          example: true
        receive_date_request:
          type: string
          format: date-time
          example: 2025-05-21T17:48:12.171Z
        status:
          $ref: '#/components/schemas/RequestRegistrationStatus'
        note:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT