# src/controllers/request.controllers.ts
import { Request, Response } from 'express'
import { ParamsDictionary } from 'express-serve-static-core'
import { REQUEST_MESSAGES } from '~/constants/messages'
import {
  ConfirmRequestProcessBloodIdReqParams,
  CreateRequestRegistrationReqBody,
  GetRequestHealthProcessByRequestIdReqParams,
  GetRequestProcessBloodIdReqParams,
  GetRequestProcessDetailIdReqParams,
  GetRequestProcessIdReqParams,
  GetRequestRegistrationIdReqParams,
  GetStatusRequestHealthProcessByRequestIdReqParams,
  UpdateRequestProcessBloodIdReqBody,
  UpdateRequestProcessBloodIdReqParams,
  UpdateRequestProcessDetailIdReqBody,
  UpdateRequestProcessDetailIdReqParams,
  UpdateRequestProcessIdReqBody,
  UpdateRequestProcessIdReqParams,
  UpdateRequestRegistrationIdReqParams,
  UpdateRequestRegistrationReqBody
} from '~/models/requests/Request.requests'
import { TokenPayload } from '~/models/requests/User.requests'
import requestsService from '~/services/request.services'

//Request - Health - Process
export const getAllRequestHealthProcessByUserIdController = async (req: Request, res: Response) => {
  const { user_id } = req.decode_authorization as TokenPayload

  const result = await requestsService.getRequestHealthProcessByUserId(user_id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_REGISTRATION_BY_USER_SUCCESS, { result })
}

export const getRequestHealthProcessByRequestIdController = async (
  req: Request<GetRequestHealthProcessByRequestIdReqParams, any, any>,
  res: Response
) => {
  const { id } = req.params

  const result = await requestsService.getRequestHealthProcessByRequestId(id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_REGISTRATION_SUCCESS, { result })
}

export const getStatusRequestHealthProcessByRequestIdController = async (
  req: Request<GetStatusRequestHealthProcessByRequestIdReqParams, any, any>,
  res: Response
) => {
  const { id } = req.params

  const result = await requestsService.getStatusRequestHealthProcessByRequestId(id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_STATUS_REQUEST_HEALTH_PROCESS_SUCCESS, { result })
}

//Request Registration
export const createRequestRegistrationController = async (
  req: Request<ParamsDictionary, any, CreateRequestRegistrationReqBody>,
  res: Response
) => {
  const { user_id } = req.decode_authorization as TokenPayload
  const { body } = req

  const result = await requestsService.createRequestRegistration({ user_id, payload: body })
  res.sendSuccess?.(REQUEST_MESSAGES.CREATE_REQUEST_REGISTRATION_SUCCESS, { result })
}

export const updateRequestRegistrationController = async (
  req: Request<UpdateRequestRegistrationIdReqParams, any, UpdateRequestRegistrationReqBody>,
  res: Response
) => {
  const { id } = req.params
  const { user_id } = req.decode_authorization as TokenPayload
  const { body } = req

  const result = await requestsService.updateRequestRegistration({ id, user_id, payload: body })
  res.sendSuccess?.(REQUEST_MESSAGES.UPDATE_REQUEST_REGISTRATION_SUCCESS, { result })
}

export const getRequestRegistrationByUserIdController = async (req: Request, res: Response) => {
  const { user_id } = req.decode_authorization as TokenPayload

  const result = await requestsService.getRequestRegistrationByUserId(user_id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_REGISTRATION_BY_USER_SUCCESS, { result })
}

export const getAllRequestRegistrationController = async (req: Request, res: Response) => {
  const result = await requestsService.getAllRequestRegistration()

  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_REGISTRATION_SUCCESS, { result })
}

export const getRequestRegistrationByIdController = async (
  req: Request<GetRequestRegistrationIdReqParams, any, any>,
  res: Response
) => {
  const { id } = req.params

  const result = await requestsService.getRequestRegistrationById(id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_REGISTRATION_SUCCESS, { result })
}

//Request Process
export const getRequestProcessByUserIdController = async (req: Request, res: Response) => {
  const { user_id } = req.decode_authorization as TokenPayload

  const result = await requestsService.getRequestProcessByUserId(user_id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_PROCESS_SUCCESS, { result })
}

export const getAllRequestProcessController = async (req: Request, res: Response) => {
  const result = await requestsService.getAllRequestProcess()
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_PROCESS_SUCCESS, { result })
}

export const getRequestProcessByIdController = async (
  req: Request<GetRequestProcessIdReqParams, any, any>,
  res: Response
) => {
  const { id } = req.params

  const result = await requestsService.getRequestProcessById(id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_PROCESS_SUCCESS, { result })
}

export const updateRequestProcessByIdController = async (
  req: Request<UpdateRequestProcessIdReqParams, any, UpdateRequestProcessIdReqBody>,
  res: Response
) => {
  const { user_id } = req.decode_authorization as TokenPayload
  const { id } = req.params
  const { body } = req

  const result = await requestsService.updateRequestProcessById({ id, user_id, payload: body })
  res.sendSuccess?.(REQUEST_MESSAGES.UPDATE_REQUEST_PROCESS_SUCCESS, { result })
}

//Request Process Detail
export const getRequestProcessDetailByProcessIdController = async (
  req: Request<GetRequestProcessDetailIdReqParams, any, any>,
  res: Response
) => {
  const { id } = req.params

  const result = await requestsService.getRequestProcessDetailByProcessId(id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_PROCESS_DETAIL_SUCCESS, { result })
}

export const updateRequestProcessDetailByProcessIdController = async (
  req: Request<UpdateRequestProcessDetailIdReqParams, any, UpdateRequestProcessDetailIdReqBody[]>,
  res: Response
) => {
  const { id } = req.params
  const { user_id } = req.decode_authorization as TokenPayload
  const { body } = req

  const result = await requestsService.updateRequestProcessDetailByProcessId({ id, user_id, payload: body })
  res.sendSuccess?.(REQUEST_MESSAGES.UPDATE_REQUEST_PROCESS_DETAIL_SUCCESS, { result })
}

//Request Process Blood
export const getRequestProcessBloodByProcessIdController = async (
  req: Request<GetRequestProcessBloodIdReqParams, any, any>,
  res: Response
) => {
  const { id } = req.params

  const result = await requestsService.getRequestProcessBloodByProcessId(id)
  res.sendSuccess?.(REQUEST_MESSAGES.GET_REQUEST_PROCESS_BLOOD_SUCCESS, { result: result })
}

export const updateRequestProcessBloodByProcessIdController = async (
  req: Request<UpdateRequestProcessBloodIdReqParams, any, UpdateRequestProcessBloodIdReqBody[]>,
  res: Response
) => {
  const { user_id } = req.decode_authorization as TokenPayload
  const { id } = req.params
  const { body } = req

  const result = await requestsService.updateRequestProcessBloodByProcessId({ id, user_id, payload: body })
  res.sendSuccess?.(REQUEST_MESSAGES.UPDATE_REQUEST_PROCESS_BLOOD_SUCCESS, { result })
}

export const confirmRequestProcessBloodByProcessIdController = async (
  req: Request<ConfirmRequestProcessBloodIdReqParams, any, any>,
  res: Response
) => {
  const { user_id } = req.decode_authorization as TokenPayload
  const { id } = req.params

  const result = await requestsService.confirmRequestProcessBloodTaken({ id, user_id })
  res.sendSuccess?.(REQUEST_MESSAGES.CONFIRM_REQUEST_PROCESS_BLOOD_SUCCESS, { result })
}