# src/controllers/user.controllers.ts
import { NextFunction, Request, Response } from 'express'
import { ParamsDictionary } from 'express-serve-static-core'
import { ObjectId } from 'mongodb'
import { HTTP_STATUS } from '~/constants/httpStatus'
import { USER_MESSAGES } from '~/constants/messages'
import { ErrorWithStatus } from '~/models/Error'
import {
  ChangeIsActiveReqParam,
  ChangePasswordReqBody,
  ChangeRoleForAdminReqBody,
  ChangeRoleReqParam,
  GetProfileByCitizenIdNumberReqParam,
  LogoutReqBody,
  RefreshTokenReqBody,
  RegisterForAdminReqBody,
  RegisterReqBody,
  TokenPayload,
  UpdateMeReqBody
} from '~/models/requests/User.requests'
import User from '~/models/schemas/User.schemas'
import usersService from '~/services/user.services'

export const loginController = async (req: Request, res: Response) => {
  const user = req.user as User
  const user_id = user._id as ObjectId
  const { fcm_token } = req.body

  // Chặn login nếu user đã bị vô hiệu hoá
  if (!user.is_active) {
    throw new ErrorWithStatus({
      message: USER_MESSAGES.USER_IS_DISABLED,
      status: HTTP_STATUS.FORBIDDEN
    })
  }

  const result = await usersService.login(user_id.toString(), fcm_token)
  res.sendSuccess?.(USER_MESSAGES.LOGIN_SUCCESS, {
    access_token: result.access_token,
    refresh_token: result.refresh_token,
    user: result.user
  })
}

export const registerController = async (
  req: Request<ParamsDictionary, any, RegisterReqBody>,
  res: Response,
  next: NextFunction
) => {
  const result = await usersService.register(req.body)
  res.sendSuccess?.(USER_MESSAGES.REGISTER_SUCCESS, {
    access_token: result.access_token,
    refresh_token: result.refresh_token
  })
}

export const logoutController = async (req: Request<ParamsDictionary, any, LogoutReqBody>, res: Response) => {
  const { refresh_token } = req.body
  const result = await usersService.logout(refresh_token)
  res.sendSuccess?.(result.message)
}

export const refreshTokenController = async (
  req: Request<ParamsDictionary, any, RefreshTokenReqBody>,
  res: Response
) => {
  const { refresh_token } = req.body
  const { user_id, role } = req.decode_refresh_token as TokenPayload

  const result = await usersService.refreshToken({ user_id, role, refresh_token })
  res.sendSuccess?.(USER_MESSAGES.REFRESH_TOKEN_SUCCESS, {
    access_token: result.access_token,
    refresh_token: result.refresh_token
  })
}

export const getMeController = async (req: Request, res: Response) => {
  const { user_id } = req.decode_authorization as TokenPayload
  const result = await usersService.getMe(user_id)
  res.sendSuccess?.(USER_MESSAGES.GET_PROFILE_SUCCESS, {
    user: result
  })
}

export const getAllUserController = async (req: Request, res: Response) => {
  const result = await usersService.getAllUsers()
  res.sendSuccess?.(USER_MESSAGES.GET_PROFILE_SUCCESS, { result })
}

export const getProfileByCitizenIdNumberController = async (
  req: Request<GetProfileByCitizenIdNumberReqParam, any, any>,
  res: Response
) => {
  const { citizen_id_number } = req.params

  const result = await usersService.getProfileByCitizenIdNumber(citizen_id_number)
  res.sendSuccess?.(USER_MESSAGES.GET_PROFILE_SUCCESS, { result })
}

export const updateMeController = async (req: Request<ParamsDictionary, any, UpdateMeReqBody>, res: Response) => {
  const { user_id } = req.decode_authorization as TokenPayload
  const { body } = req

  const result = await usersService.updateMe(user_id, body)
  res.sendSuccess?.(USER_MESSAGES.UPDATE_PROFILE_SUCCESS, { result })
}

export const changePasswordController = async (
  req: Request<ParamsDictionary, any, ChangePasswordReqBody>,
  res: Response
) => {
  const { user_id } = req.decode_authorization as TokenPayload
  const { password } = req.body

  const result = await usersService.changePassword(user_id, password)
  res.sendSuccess?.(result.message)
}

export const changeIsActiveController = async (req: Request<ChangeIsActiveReqParam, any, any>, res: Response) => {
  const { user_id } = req.params

  const result = await usersService.changeIsActive(user_id)
  res.sendSuccess?.(result.message)
}

export const registerForAdminController = async (
  req: Request<ParamsDictionary, any, RegisterForAdminReqBody>,
  res: Response
) => {
  const result = await usersService.registerForAdmin(req.body)
  res.sendSuccess?.(USER_MESSAGES.REGISTER_SUCCESS, {
    result
  })
}

export const changeRoleForAdminController = async (
  req: Request<ChangeRoleReqParam, any, ChangeRoleForAdminReqBody>,
  res: Response
) => {
  const { user_id } = req.params
  const { body } = req

  const result = await usersService.changeRoleForAdmin(user_id, body)
  res.sendSuccess?.(result.message)
}